+++ id3lib/src/tag_parse_v1.cpp
@@ -1,4 +1,4 @@
-// $Id: tag_parse_v1.cpp,v 1.1 2006-01-17 21:09:53 stulleamgym Exp $
+// $Id: tag_parse_v1.cpp,v 1.2 2008-10-07 17:20:33 stulleamgym Exp $
 
 // id3lib: a C++ library for creating and manipulating id3v1/v2 tags
 // Copyright 1999, 2000  Scott Thomas Haug
@@ -115,7 +115,7 @@ bool id3::v1::parse(ID3_TagImpl& tag, ID3_Reader& reader)
   {
     if (trackno[1] != '\0')
     { //we've got a tracknumber
-      size_t track = trackno[1];
+      uchar track = trackno[1];
       field = id3::v2::getTrack(tag);
       if (field.size() == 0 || field == "00")
       {
@@ -128,7 +128,7 @@ bool id3::v1::parse(ID3_TagImpl& tag, ID3_Reader& reader)
   else
   {
     // trackno[0] != '\0'
-    const int paddingsize = (ID3_V1_LEN_COMMENT-2) - comment.size();
+    const size_t paddingsize = (ID3_V1_LEN_COMMENT-2) - comment.size();
     const char * padding = "                            "; //28 spaces
 
     if (trackno[1] == '\0' || trackno[1] == 0x20 && trackno[0] != 0x20)
@@ -152,7 +152,7 @@ bool id3::v1::parse(ID3_TagImpl& tag, ID3_Reader& reader)
   
   ID3D_NOTICE("id3::v1::parse: read bytes: " << reader.getCur() - beg);
   // the GENRE field/frame
-  uchar genre = reader.readChar();
+  uchar genre = static_cast<uchar>(reader.readChar());
   field = id3::v2::getGenre(tag);
   if (genre != 0xFF && (field.size() == 0 || field == ""))
   {
